CSS Pruning, Implementation of Download Manager, Download Tracking
Due to the fact a good percentage of the codebase of CSS due to hacking together the application and not making a sure fire decision of what UI framework we'll use (which is officially react-materialize).Also the front-end of the Download Manager has been implemented with temporary code and examples until the back-end is integrated and basic tests have been built for it. Finally we've begun implementing the sqlite database but for temporary storage and file tracking until a user upgrades or chooses to empty it(soon to be with a click of a button).

Signed-off-by: Kthulu120 <Troy.L.Smith-1@ou.edu>