Youtube-DL subscriptions is now live and working
Youtube-dl subscriptions is now working though only on initial start, the downloads do not auto resume nor can you restart them from the GUI though it is planned. Also Have not finished adding in all providers though they should be added in the next commit. Along with this the database is now funcitonal for downloads and subscriptions.Errors will continue to be implemented for users to see.

Signed-off-by: Kthulu120 <Troy.L.Smith-1@ou.edu>